requirements.txt
# Airflow image already contains Airflow; we only add libs we need.
pyspark==3.5.1
psycopg2-binary==2.9.9
pandas==2.2.2
SQLAlchemy<2.0
pyyaml==6.0.2
